requirements.txt
discord.py[voice]>=2.4.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0
mutagen>=1.47.0
aiosqlite>=0.19.0
